% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfbm.R
\name{?}
\alias{?}
\title{Redifined question mark for benchmarking}
\usage{
"?"(e1, e2)
}
\arguments{
\item{e2}{}
}
\description{
It evaluates its rhs so the function still works when edited with `call1 ? call2` syntax
}
