% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfbm.R
\name{selfbm}
\alias{selfbm}
\title{Benchmark a function against itself}
\usage{
selfbm(call, ..., engine = c("microbenchmark", "bench"))
}
\arguments{
\item{call}{a function call}

\item{...}{additional parameters passed to the relevant engine}

\item{engine}{engine}
}
\description{
Benchmark a function against itself
}
